% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggintervals.R
\name{bin_int}
\alias{bin_int}
\title{creates a confidence level for the mean of independent binary data, using the
exact test (which yields intervals that are a little conservative for small
sample sizes, but guaranteed to capture >= 0.95 probability)}
\usage{
bin_int(x, conf.int = 0.95)
}
\arguments{
\item{x}{a vector of (numeric) binary 0s and 1s}

\item{conf.int}{the size of the confidence interval}
}
\description{
creates a confidence level for the mean of independent binary data, using the
exact test (which yields intervals that are a little conservative for small
sample sizes, but guaranteed to capture >= 0.95 probability)
}

