% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ggintervals}
\alias{ggintervals}
\alias{ggintervals-package}
\title{ggintervals: Additional useful confidence intervals to use with ggplot2.}
\description{
ggintervals: Additional useful confidence intervals to use with ggplot2.
}

