% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggintervals.R
\name{mean_cl_boot_bca}
\alias{mean_cl_boot_bca}
\title{creates a bootstrapped interval similar to ggplot2's mean_cl_boot, but uses
the better BCA bootstrap method}
\usage{
mean_cl_boot_bca(x, conf.int = 0.95, replicates = 1000)
}
\arguments{
\item{x}{a vector of values to bootstrap a mean from}

\item{conf.int}{the size of the bootstrapped confidence interval}

\item{replicates}{the number of bootstrap replicates}
}
\value{
a data frame in a format suitable for plotting with ggplot2
  containing the bootstrapped mean and lower and upper bounds
}
\description{
creates a bootstrapped interval similar to ggplot2's mean_cl_boot, but uses
the better BCA bootstrap method
}

